+++ testsuite/grid/test_range_fortran_iteration.cpp
@@ -32,7 +32,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_1d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<1>(lo, hi);
-        Range<int, 1, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 1, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
 
         int count = 0;
@@ -40,7 +40,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_1d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
             BOOST_CHECK_EQUAL(grid(i), i - lo[0]);
         }
@@ -59,7 +59,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_2d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<2>(lo, hi);
-        Range<int, 2, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 2, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -68,9 +68,9 @@ BOOST_FIXTURE_TEST_CASE( iterate_2d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
                 BOOST_CHECK_EQUAL(grid(i, j), i - lo[0] + (j-lo[1])*dims[0]);
             }
@@ -90,7 +90,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_3d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<3>(lo, hi);
-        Range<int, 3, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 3, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -99,11 +99,11 @@ BOOST_FIXTURE_TEST_CASE( iterate_3d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
                     BOOST_CHECK_EQUAL(grid(i, j, k), i - lo[0] + (j-lo[1] + (k - lo[2])*dims[1])*dims[0]);
                 }
@@ -124,7 +124,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_4d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<4>(lo, hi);
-        Range<int, 4, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 4, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -133,13 +133,13 @@ BOOST_FIXTURE_TEST_CASE( iterate_4d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
                         BOOST_CHECK_EQUAL(grid(i, j, k, l), 
                             i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3])*dims[2])*dims[1])*dims[0]
@@ -164,7 +164,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_5d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<5>(lo, hi);
-        Range<int, 5, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 5, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -173,15 +173,15 @@ BOOST_FIXTURE_TEST_CASE( iterate_5d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
                             BOOST_CHECK_EQUAL(grid(i, j, k, l, m), 
                                 i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3]
@@ -209,7 +209,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_6d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<6>(lo, hi);
-        Range<int, 6, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 6, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -218,17 +218,17 @@ BOOST_FIXTURE_TEST_CASE( iterate_6d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
-                            for (int n=lo[5]; n<=hi[5]; ++n)
+                            for (ptrdiff_t n=lo[5]; n<=hi[5]; ++n)
                             {
                                 BOOST_CHECK_EQUAL(grid(i, j, k, l, m, n), 
                                     i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3]
@@ -256,7 +256,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_7d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<7>(lo, hi);
-        Range<int, 7, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 7, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -265,19 +265,19 @@ BOOST_FIXTURE_TEST_CASE( iterate_7d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
-                            for (int n=lo[5]; n<=hi[5]; ++n)
+                            for (ptrdiff_t n=lo[5]; n<=hi[5]; ++n)
                             {
-                                for (int o=lo[6]; o<=hi[6]; ++o)
+                                for (ptrdiff_t o=lo[6]; o<=hi[6]; ++o)
                                 {
                                     BOOST_CHECK_EQUAL(grid(i, j, k, l, m, n, o), 
                                         i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3]
@@ -307,7 +307,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_8d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<8>(lo, hi);
-        Range<int, 8, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 8, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -316,21 +316,21 @@ BOOST_FIXTURE_TEST_CASE( iterate_8d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
-                            for (int n=lo[5]; n<=hi[5]; ++n)
+                            for (ptrdiff_t n=lo[5]; n<=hi[5]; ++n)
                             {
-                                for (int o=lo[6]; o<=hi[6]; ++o)
+                                for (ptrdiff_t o=lo[6]; o<=hi[6]; ++o)
                                 {
-                                    for (int p=lo[7]; p<=hi[7]; ++p)
+                                    for (ptrdiff_t p=lo[7]; p<=hi[7]; ++p)
                                     {
                                         BOOST_CHECK_EQUAL(grid(i, j, k, l, m, n, o, p), 
                                             i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3]
@@ -361,7 +361,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_9d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<9>(lo, hi);
-        Range<int, 9, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 9, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -370,23 +370,23 @@ BOOST_FIXTURE_TEST_CASE( iterate_9d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
-                            for (int n=lo[5]; n<=hi[5]; ++n)
+                            for (ptrdiff_t n=lo[5]; n<=hi[5]; ++n)
                             {
-                                for (int o=lo[6]; o<=hi[6]; ++o)
+                                for (ptrdiff_t o=lo[6]; o<=hi[6]; ++o)
                                 {
-                                    for (int p=lo[7]; p<=hi[7]; ++p)
+                                    for (ptrdiff_t p=lo[7]; p<=hi[7]; ++p)
                                     {
-                                        for (int q=lo[8]; q<=hi[8]; ++q)
+                                        for (ptrdiff_t q=lo[8]; q<=hi[8]; ++q)
                                         {
                                             BOOST_CHECK_EQUAL(grid(i, j, k, l, m, n, o, p, q), 
                                                 i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3] + (m - lo[4] 
@@ -417,7 +417,7 @@ BOOST_FIXTURE_TEST_CASE( iterate_10d, RangeIterationTest )
     for (int n=0; n<10; ++n)
     {
         random_extent<10>(lo, hi);
-        Range<int, 10, ArrayBoostTestArgCheck> range(lo, hi);
+        Range<ptrdiff_t, 10, ArrayBoostTestArgCheck> range(lo, hi);
         GridType grid(lo, hi);
         auto dims = grid.getDims();
 
@@ -426,25 +426,25 @@ BOOST_FIXTURE_TEST_CASE( iterate_10d, RangeIterationTest )
             grid[pos] = count++;
         });
 
-        for (int i=lo[0]; i<=hi[0]; ++i)
+        for (ptrdiff_t i=lo[0]; i<=hi[0]; ++i)
         {
-            for (int j=lo[1]; j<=hi[1]; ++j)
+            for (ptrdiff_t j=lo[1]; j<=hi[1]; ++j)
             {
-                for (int k=lo[2]; k<=hi[2]; ++k)
+                for (ptrdiff_t k=lo[2]; k<=hi[2]; ++k)
                 {
-                    for (int l=lo[3]; l<=hi[3]; ++l)
+                    for (ptrdiff_t l=lo[3]; l<=hi[3]; ++l)
                     {
-                        for (int m=lo[4]; m<=hi[4]; ++m)
+                        for (ptrdiff_t m=lo[4]; m<=hi[4]; ++m)
                         {
-                            for (int n=lo[5]; n<=hi[5]; ++n)
+                            for (ptrdiff_t n=lo[5]; n<=hi[5]; ++n)
                             {
-                                for (int o=lo[6]; o<=hi[6]; ++o)
+                                for (ptrdiff_t o=lo[6]; o<=hi[6]; ++o)
                                 {
-                                    for (int p=lo[7]; p<=hi[7]; ++p)
+                                    for (ptrdiff_t p=lo[7]; p<=hi[7]; ++p)
                                     {
-                                        for (int q=lo[8]; q<=hi[8]; ++q)
+                                        for (ptrdiff_t q=lo[8]; q<=hi[8]; ++q)
                                         {
-                                            for (int r=lo[9]; r<=hi[9]; ++r)
+                                            for (ptrdiff_t r=lo[9]; r<=hi[9]; ++r)
                                             {
                                                 BOOST_CHECK_EQUAL(grid(i, j, k, l, m, n, o, p, q, r), 
                                                     i - lo[0] + (j-lo[1] + (k - lo[2] + (l - lo[3] + (m - lo[4]
